refactor: place enums and types into folders
