+++ news-JS/src/components/controller/Loader/enums/ApiMethods.ts
@@ -1,0 +1,6 @@
+enum ApiMethods {
+  GET = 'GET',
+  POST = 'POST',
+}
+
+export default ApiMethods;
