+++ news-JS/src/components/controller/Loader/enums/Status.ts
@@ -1,0 +1,6 @@
+enum Status {
+  Unauthorized = 401,
+  NotFound = 404,
+}
+
+export default Status;
