+++ news-JS/src/components/controller/Loader/types/ApiKey.ts
@@ -1,0 +1,5 @@
+type ApiKey = {
+  [apiKey: string]: string;
+};
+
+export default ApiKey;
