+++ news-JS/src/components/controller/Loader/types/Endpoints.ts
@@ -1,0 +1,3 @@
+type Endpoints = 'sources' | 'everything';
+
+export default Endpoints;
