+++ news-JS/src/components/controller/Loader/types/Options.ts
@@ -1,0 +1,5 @@
+type Options = {
+  [sources: string]: string;
+};
+
+export default Options;
