+++ news-JS/src/components/controller/Loader/types/Responce.ts
@@ -1,0 +1,9 @@
+import Options from './Options';
+import Endpoints from './Endpoints';
+
+type Responce = {
+  endpoint: Endpoints;
+  options?: Partial<Options>;
+};
+
+export default Responce;
