+++ news-JS/src/components/interfaces/IDataNews.ts
@@ -1,5 +1,9 @@
+import IDataArticle from './IDataArticle';
+
 interface IDataNews {
   status: 'ok' | 'error';
   totalResults: number;
   articles: Array<IDataArticle>;
 }
+
+export default IDataNews;
