+++ news-JS/src/components/interfaces/IDataSource.ts
@@ -6,3 +6,5 @@ interface IDataSource {
   name: string;
   url: string;
 }
+
+export default IDataSource;
