+++ news-JS/src/components/interfaces/IDataSources.ts
@@ -1,4 +1,8 @@
+import IDataSource from './IDataSource';
+
 interface IDataSources {
   status: 'ok' | 'error';
   sources: Array<IDataSource>;
 }
+
+export default IDataSources;
